TermProject2/requirements.txt
torch~=2.0.1
torchvision~=0.15.2
matplotlib~=3.8.1
torchcam~=0.4.0
torchviz~=0.0.2
numpy~=1.25.2
tqdm~=4.66.1
scikit-learn~=1.3.1
pandas~=2.1.4
pillow~=10.0.0